Fix displaying error image via snapshot when stream is lost
The frame buffer used to display a snapshot of the stream was not flushed when the connection to the stream was lost and the relay transitioned to displaying the error image. This caused browsers to complain of a content length mismatch due to partial frames remaining in the frame buffer.