Simplify HTTP client joining and streaming process
Standard clients joining via HTTP requests were introduced to the stream in an unnecessarily convoluted way, which involved sending the current frame before it was fully received from the stream source. This process was of questionable benefit, and has been replaced with a simplified process which closely matches that of the WebSocket clients, in which the frame is only sent when fully received from the stream source.